+++ tests/Tests/Parametizer/Autocompletion/scripts/different-params.php
@@ -1,30 +1,7 @@
 <?php declare(strict_types=1);
 
-use MagicPush\CliToolkit\Parametizer\Parametizer;
+use MagicPush\CliToolkit\Tests\Tests\Parametizer\Autocompletion\Config\DifferentParams;
 
 require_once __DIR__ . '/../../../init-console.php';
 
-Parametizer::newConfig()
-    ->newOption('--opt', '-o')
-    ->allowedValuesDescribed([
-        '100'     => 'One hundred',
-        '200'     => 'Two hundred',
-        '1000000' => 'One million',
-    ])
-
-    ->newOption('--any-value', '-a')
-
-    ->newFlag('--flag', '-f')
-    ->newFlag('--second-flag', '-s')
-
-    ->newArgument('arg')
-    ->allowedValues(['super', 'prefix', 'premium'])
-
-    ->newArgument('arg-allowed-values')
-    ->allowedValues(['aaa', 'bbb', 'ccc'])
-
-    ->newArgument('arg-allowed-values-empty')
-    ->allowedValues(['aaa', 'bbb', 'ccc'])
-    ->allowedValues([]) // Overwrite allowed values and ensure that previous completion list is emptied.
-
-    ->run();
+DifferentParams::getConfigBuilder()->run();